refactor: replace legacy implementation with modern ES6+ version
- Replace lib/btree.js with modern ES6+ implementation
- Keep legacy version as lib/btree-legacy.js for reference
- Update index.js to use modern implementation as default
- Maintain full backward compatibility
- All existing APIs work exactly the same
- Modern features: classes, arrow functions, template literals, destructuring
- Improved error handling and validation
- Better code organization and documentation

Breaking changes: None (fully backward compatible)
New features: BTree class, size(), isEmpty(), clear() methods
Modern syntax: ES6+ classes, const/let, arrow functions, template literals
