+++ index.js
@@ -1,15 +1,4 @@
-// Export both modern and legacy versions for compatibility
-const modernBTree = require('./lib/btree-modern');
-const legacyBTree = require('./lib/btree');
+// Export modern ES6+ B-tree implementation
+const btree = require('./lib/btree');
 
-module.exports = {
-    // Modern ES6+ class
-    BTree: modernBTree.BTree,
-    createTree: modernBTree.createTree,
-    
-    // Legacy support
-    legacy: legacyBTree,
-    
-    // Default export (modern)
-    ...modernBTree
-};
+module.exports = btree;
